Calculate next() from that point in time
Timezone conversion was being done with startDate, instead of result. That meant that the timezone offset was being applied to the first instance, not the occurrence date.

In other words, it'd show times like:

..during DST..
08:45am
..DST ends..
07:45am

Which was confusing because the bell would still run at 8:45am on that day, but looking ahead, that wasn't obvious.

This still needs testing, but early signs show that it'll work.